Implement projectile creation and management
Added the ability to generate and manage projectiles using the newly created ProjectileFactory. Additionally, a new "Projectile" component has been introduced to store the projectile direction. In-game, the player can now create a projectile by clicking the MOUSE_LEFT_BUTTON. The projectiles' movement is managed and they are removed if they move off-screen.